Create new utils file for confserver
