+++ curiefense/curieconf/server/curieconf/confserver/v1/api.py
@@ -9,6 +9,7 @@ from collections import defaultdict
 import datetime
 from curieconf import utils
 from curieconf.utils import cloud
+from curieconf.confserver import utils as apiutils
 import requests
 from jsonschema import validate
 from pathlib import Path
@@ -256,7 +257,7 @@ m_document_list_entry = api.model(
 
 m_config_documents = api.model(
     "Config Documents",
-    {x: fields.List(fields.Nested(models[x], default=[]), attribute=utils.vconvert(x, "v1")) for x in models},
+    {x: fields.List(fields.Nested(models[x], default=[]), attribute=apiutils.vconvert(x, "v1")) for x in models},
 )
 
 m_config_blobs = api.model(
@@ -513,7 +514,7 @@ class DocumentsResource(Resource):
         res = current_app.backend.documents_list(config)
         # convert to v1 names
         for doc in res:
-            doc['name'] = utils.vconvert(doc['name'], "v1", True)
+            doc['name'] = apiutils.vconvert(doc['name'], "v1", True)
         return res
 
 
@@ -524,16 +525,16 @@ class DocumentResource(Resource):
         "Get a complete document"
         if document not in models:
             abort(404, "document does not exist")
-        res = current_app.backend.documents_get(config, utils.vconvert(document, "v1"))
+        res = current_app.backend.documents_get(config, apiutils.vconvert(document, "v1"))
         return marshal(res, models[document], skip_none=True)
 
     def post(self, config, document):
         "Create a new complete document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, apiutils.model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.documents_create(
-            config, utils.vconvert(document, "v1"), data
+            config, apiutils.vconvert(document, "v1"), data
         )
         return res
 
@@ -541,9 +542,9 @@ class DocumentResource(Resource):
         "Update an existing document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, apiutils.model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.documents_update(
-            config, utils.vconvert(document, "v1"), data
+            config, apiutils.vconvert(document, "v1"), data
         )
         return res
 
@@ -552,7 +553,7 @@ class DocumentResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.documents_delete(
-            config, utils.vconvert(document, "v1")
+            config, apiutils.vconvert(document, "v1")
         )
         return res
 
@@ -564,7 +565,7 @@ class DocumentListVersionResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.documents_list_versions(
-            config, utils.vconvert(document, "v1")
+            config, apiutils.vconvert(document, "v1")
         )
         return marshal(res, m_version_log, skip_none=True)
 
@@ -576,7 +577,7 @@ class DocumentVersionResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.documents_get(
-            config, utils.vconvert(document, "v1"), version
+            config, apiutils.vconvert(document, "v1"), version
         )
         return marshal(res, models[document], skip_none=True)
 
@@ -586,7 +587,7 @@ class DocumentRevertResource(Resource):
     def put(self, config, document, version):
         "Create a new version for a document from an old version"
         return current_app.backend.documents_revert(
-            config, utils.vconvert(document, "v1"), version
+            config, apiutils.vconvert(document, "v1"), version
         )
 
 
@@ -601,16 +602,16 @@ class EntriesResource(Resource):
         "Retrieve the list of entries in a document"
         if document not in models:
             abort(404, "document does not exist")
-        res = current_app.backend.entries_list(config, utils.vconvert(document, "v1"))
+        res = current_app.backend.entries_list(config, apiutils.vconvert(document, "v1"))
         return res  # XXX: marshal
 
     def post(self, config, document):
         "Create an entry in a document"
         if document not in models:
             abort(404, "document does not exist")
-        data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+        data = marshal(request.json, apiutils.model_invert_names(models[document]), skip_none=True)
         res = current_app.backend.entries_create(
-            config, utils.vconvert(document, "v1"), data
+            config, apiutils.vconvert(document, "v1"), data
         )
         return res
 
@@ -622,7 +623,7 @@ class EntryResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.entries_get(
-            config, utils.vconvert(document, "v1"), entry
+            config, apiutils.vconvert(document, "v1"), entry
         )
         return marshal(res, models[document], skip_none=True)
 
@@ -632,9 +633,9 @@ class EntryResource(Resource):
             abort(404, "document does not exist")
         isValid = validateJson(request.json, document)
         if isValid:
-            data = marshal(request.json, utils.model_invert_names(models[document]), skip_none=True)
+            data = marshal(request.json, apiutils.model_invert_names(models[document]), skip_none=True)
             res = current_app.backend.entries_update(
-                config, utils.vconvert(document, "v1"), entry, data
+                config, apiutils.vconvert(document, "v1"), entry, data
             )
             return res
         else:
@@ -645,7 +646,7 @@ class EntryResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.entries_delete(
-            config, utils.vconvert(document, "v1"), entry
+            config, apiutils.vconvert(document, "v1"), entry
         )
         return res
 
@@ -664,9 +665,9 @@ class EntryEditResource(Resource):
         for edit in data:
             mapped = pydash.objects.set_({}, edit['path'], edit['value'])
             marshaled_map = marshal(mapped, models[document], skip_none=True)
-            utils.dict_to_path_value(marshaled_map, starting_path_list=converted_names_data)
+            apiutils.dict_to_path_value(marshaled_map, starting_path_list=converted_names_data)
         res = current_app.backend.entries_edit(
-            config, utils.vconvert(document, "v1"), entry, data
+            config, apiutils.vconvert(document, "v1"), entry, data
         )
         return res
 
@@ -678,7 +679,7 @@ class EntryListVersionResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.entries_list_versions(
-            config, utils.vconvert(document, "v1"), entry
+            config, apiutils.vconvert(document, "v1"), entry
         )
         return marshal(res, m_version_log, skip_none=True)
 
@@ -692,7 +693,7 @@ class EntryVersionResource(Resource):
         if document not in models:
             abort(404, "document does not exist")
         res = current_app.backend.entries_get(
-            config, utils.vconvert(document, "v1"), entry, version
+            config, apiutils.vconvert(document, "v1"), entry, version
         )
         return marshal(res, models[document], skip_none=True)
