+++ curiefense/curieconf/test/test_utils.py
@@ -4,6 +4,7 @@ import json
 import codecs
 import base64
 from flask_restplus import fields, model
+from curieconf.confserver import utils as apiutils
 
 binvec_hex = (
     "b70a1da09a4998bd56b083d76bf528053c9b924bbb07168792151a5a177bbaa232949a8600bcb2"
@@ -70,32 +71,32 @@ def test_bytes2jblob_json():
     assert res2 == vec
 
 def test_vconvert():
-    assert utils.vconvert("urlmaps", "v1", False) == "securitypolicies"
-    assert utils.vconvert("wafrules", "v1", False) == "contentfilterrules"
-    assert utils.vconvert("wafpolicies", "v1", False) == "contentfilterprofiles"
-    assert utils.vconvert("aclpolicies", "v1", False) == "aclprofiles"
-    assert utils.vconvert("tagrules", "v1", False) == "globalfilters"
-    assert utils.vconvert("flowcontrol", "v1", False) == "flowcontrolpolicies"
-    assert utils.vconvert("securitypolicies", "v1", True) == "urlmaps"
-    assert utils.vconvert("contentfilterrules", "v1", True) == "wafrules"
-    assert utils.vconvert("contentfilterprofiles", "v1", True) == "wafpolicies"
-    assert utils.vconvert("aclprofiles", "v1", True) == "aclpolicies"
-    assert utils.vconvert("globalfilters", "v1", True) == "tagrules"
-    assert utils.vconvert("flowcontrolpolicies", "v1", True) == "flowcontrol"
-    assert utils.vconvert("something", "v1", False) == "something"
+    assert apiutils.vconvert("urlmaps", "v1", False) == "securitypolicies"
+    assert apiutils.vconvert("wafrules", "v1", False) == "contentfilterrules"
+    assert apiutils.vconvert("wafpolicies", "v1", False) == "contentfilterprofiles"
+    assert apiutils.vconvert("aclpolicies", "v1", False) == "aclprofiles"
+    assert apiutils.vconvert("tagrules", "v1", False) == "globalfilters"
+    assert apiutils.vconvert("flowcontrol", "v1", False) == "flowcontrolpolicies"
+    assert apiutils.vconvert("securitypolicies", "v1", True) == "urlmaps"
+    assert apiutils.vconvert("contentfilterrules", "v1", True) == "wafrules"
+    assert apiutils.vconvert("contentfilterprofiles", "v1", True) == "wafpolicies"
+    assert apiutils.vconvert("aclprofiles", "v1", True) == "aclpolicies"
+    assert apiutils.vconvert("globalfilters", "v1", True) == "tagrules"
+    assert apiutils.vconvert("flowcontrolpolicies", "v1", True) == "flowcontrol"
+    assert apiutils.vconvert("something", "v1", False) == "something"
 
 def test_model_invert_names():
     mod1 = model.Model("test", {
             "test":fields.String(attribute="test2")
     })
-    res = utils.model_invert_names(mod1)
+    res = apiutils.model_invert_names(mod1)
     assert res.name == mod1.name and type(res['test2']) is fields.String \
         and res['test2'].attribute == 'test'
 
     mod2 = model.Model("test", {
             "test":fields.Nested(mod1, attribute="test2")
     })
-    res = utils.model_invert_names(mod2)
+    res = apiutils.model_invert_names(mod2)
     assert res.name == mod2.name \
         and type(res['test2']) is fields.Nested \
         and res['test2'].attribute == 'test' \
@@ -105,7 +106,7 @@ def test_model_invert_names():
     mod3 = model.Model("test", {
             "test":fields.List(mod1, attribute="test2")
     })
-    res = utils.model_invert_names(mod3)
+    res = apiutils.model_invert_names(mod3)
     assert res.name == mod3.name \
         and type(res['test2']) is fields.List \
         and res['test2'].attribute == 'test' \
@@ -115,7 +116,7 @@ def test_model_invert_names():
     mod4 = model.Model("test", {
             "test*":fields.Wildcard(mod2, attribute="test2*")
     })
-    res = utils.model_invert_names(mod4)
+    res = apiutils.model_invert_names(mod4)
     assert res.name == mod4.name \
         and type(res['test2']) is fields.Wildcard \
         and res['test2'].attribute == 'test' \
@@ -125,7 +126,7 @@ def test_model_invert_names():
         and res['test2'].container['test2'].model['test2'].attribute == 'test'
 
 def test_dict_to_path_value():
-    assert utils.dict_to_path_value({}) == []
-    assert utils.dict_to_path_value({'a':1,'b':{'b':2}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b', 'value': 2}]
-    assert utils.dict_to_path_value({'a':1,'b':{'b':{'b': 2}}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b.b', 'value': 2}]
-    assert utils.dict_to_path_value({'a':1,'b':{'b':2}, 'c':{'c':{'c': 3}}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b', 'value': 2},  {'path': 'c.c.c', 'value': 3}]
+    assert apiutils.dict_to_path_value({}) == []
+    assert apiutils.dict_to_path_value({'a':1,'b':{'b':2}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b', 'value': 2}]
+    assert apiutils.dict_to_path_value({'a':1,'b':{'b':{'b': 2}}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b.b', 'value': 2}]
+    assert apiutils.dict_to_path_value({'a':1,'b':{'b':2}, 'c':{'c':{'c': 3}}}) == [{'path': 'a', 'value': 1}, {'path': 'b.b', 'value': 2},  {'path': 'c.c.c', 'value': 3}]
